fix(test): add `node 6` support for tests
The version of `lab` stated in the devDependencies did not support `node 6`. Running the tests would throw an error of `The following leaks were detected:Reflect`. On `lab 10.3.2`, `node 6` support is added and `Reflect` is added to the globals whitelist, fixing that error (see https://github.com/hapijs/lab/issues/568).

Upgrading `lab` and `code` to their latest major versions.